[Tweak] Better Telescopic Baton
