+++ Content.Shared/_White/Item/Telebaton/TelescopicBatonComponent.cs
@@ -1,0 +1,6 @@
+namespace Content.Shared._White.Item.TelescopicBaton;
+
+[RegisterComponent]
+public sealed partial class TelescopicBatonComponent : Component
+{
+}
