+++ Resources/Locale/ru-RU/_white/object/telebaton.ftl
@@ -1,4 +1,4 @@
-ent-Telebaton = телескопическая дубинка
+ent-TelescopicBaton = телескопическая дубинка
     .desc = Компактное, но надежное оружие личной обороны. В сложенном состоянии может быть скрыто.
 
 comp-telebaton-examined-on = Дубинка в боевом положении.
