+++ Resources/Prototypes/Catalog/Fills/Backpacks/StarterGear/backpack.yml
@@ -66,9 +66,9 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
+      - id: TelescopicBaton # WD
       #- name: StationCharter
-      #- name: TelescopicBaton
+
 - type: entity
   noSpawn: true
   parent: ClothingBackpackEngineering
@@ -78,8 +78,7 @@
     contents:
       - id: BoxSurvivalEngineering
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -90,8 +89,7 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -102,8 +100,7 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -114,8 +111,7 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -126,8 +122,7 @@
     contents:
       - id: BoxSurvivalMedical
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -138,8 +133,7 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
