+++ Resources/Prototypes/Catalog/Fills/Backpacks/StarterGear/duffelbag.yml
@@ -75,9 +75,9 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
+      - id: TelescopicBaton # WD
       #- name: StationCharter
-      #- name: TelescopicBaton
+
 - type: entity
   noSpawn: true
   parent: ClothingBackpackDuffelEngineering
@@ -87,8 +87,7 @@
     contents:
       - id: BoxSurvivalEngineering
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -99,8 +98,7 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -111,8 +109,7 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -123,8 +120,7 @@
     contents:
       - id: BoxSurvivalMedical
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
@@ -135,8 +131,7 @@
     contents:
       - id: BoxSurvival
       - id: Flash
-      - id: Telebaton
-      #- id: TelescopicBaton
+      - id: TelescopicBaton # WD
 
 - type: entity
   noSpawn: true
